requirements.txt
asgiref==3.3.1
asn1crypto==0.24.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.11.5
chardet==3.0.4
configobj==5.0.6
cryptography==2.3
dbus-python==1.2.4
decorator==4.2.1
Django==3.1.6
django-markdown==0.8.4
django-mdeditor==0.1.18
django-pure-pagination==0.3.0
django-simpleui==2021.2
gpg==1.10.0
idna==2.10
importlib-metadata==1.7.0
isc==2.0
Jinja2==2.11.2
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
netifaces==0.10.6
nftables==0.1
oauthlib==3.1.0
pciutils==2.3.6
perf==0.1
ply==3.9
pycairo==1.16.3
pycparser==2.14
pygobject==3.28.3
pyOpenSSL==18.0.0
pyrsistent==0.17.3
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
pytz==2021.1
pyudev==0.21.0
PyYAML==5.3.1
requests==2.24.0
rhnlib==2.8.6
rpm==4.14.2
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.2.2
setroubleshoot==1.1
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sqlparse==0.4.1
systemd-python==234
urllib3==1.25.10
uWSGI==2.0.19.1
zipp==3.1.0
